Update Deep Layout Parsing.ipynb
The current version of the Deep Learning Parsing example Notebook lacks the 'inplace=True' parameter on the sort method call to the Layout instance in the left Interval. This version's large block of image data shows these TextBlock's ids in top-down sequential order. Running the Notebook as written, however, will refresh the Notebook page image with out of sequence ids due to the default of the inplace parameter being False.

For instructional clarity, this version casts the right_blocks variable as a Layout instance initialized with the TextBlocks not in the left Interval and with inplace=True.

A few non-essential additions to the source comments are added relative to the changes described.